Renamed classes for consistency and moved to individual files.
